Moved deflectometry/lib/Geometry... to sofast
