+++ opencsp/app/sofast/lib/ParamsSofastFixed.py
@@ -1,8 +1,8 @@
 """Parameters class for FixedPatternScreen class"""
-from opencsp.common.lib.deflectometry.GeometryProcessingParams import (
+from opencsp.app.sofast.lib.GeometryProcessingParams import (
     GeometryProcessingParams,
 )
-from opencsp.common.lib.deflectometry.GeometryDataDebug import GeometryDataDebug
+from opencsp.app.sofast.lib.GeometryDataDebug import GeometryDataDebug
 from opencsp.common.lib.deflectometry.SlopeSolverDataDebug import SlopeSolverDataDebug
